Project-05_Custom_LangChain_AI_Assistant/requirements.txt
python-dotenv>=1.0.0
gradio>=4.0.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0